Do not use function pointers of unspecified arguments
This is necessary in order to build with C23, which is the default
standard in GCC 15.

In C23, `int foo()` is no longer a function taking unspecified
arguments, but a function taking no arguments, like `int foo(void)`.

Furthermore, treating a pointer to a function of unspecified arguments
as compatible with a pointer to a function with declared arguments was
probably always dubious, although it worked in practice.